ch4/ucx: defined common routines to check valid win/target.
UCX enables RMA only for window with create/allocate flavors. PROC_NULL
target or target with zero window size is also not handled in UCX. Thus,
we define common checking routines to check if window or target is
reachable in UCX. The checking routines used in both RMA issuing and
sync hooks.