shm: fix bug in shmem segment allocation mechanism
The `MPIDU_shm_seg_alloc` and `MPIDU_shm_seg_commit` book memory space
for shared memory and allocate it, respectively. The amount of the total
memory requested is tracked by the alloc function using a static global
variable (i.e., `segment_len`). When the commit terminates this variable
should be reset to zero but it's not. This does not seem to be a problem
in CH3. However, in CH4 after allocating the first shared memory segment
for intranode communication, the netmods themselves allocate another
segment for business card exchange. Because `segment_len` is never reset
the length of such segment adds to the previous intranode shmem segment
length.

This patch fixes the problem by resetting `segment_len` to zero at the
end of the commit function.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>