ch4/ofi: Close per-window counter on failure path
When creating a window with scalable endpoints or shared transmit
contexts, there is a case where first per-window counter is
allocated but later some other initialization fails.
In such a case this counter is going to leak.
This patch fixes it by properly closing the counter in case of
failure.

Signed-off-by: Min Si <msi@anl.gov>