test : assert to ensure context shift bits are non zero
This is an update to check that rma window creation fails
with ch4 ofi netmod. Window creation should fail because
max_rma_key_bits is set to a value less than context_shift
count, resulting in a negative count for shift value
used to compute the number of maximum allowed contexts.
The bit shift value is computed to be -1 currently, and left
shifting by a negative value behaves differently on each
platform, and cannot be relied on.

Signed-off-by: Pavan Balaji <balaji@anl.gov>