ch4/generic: Correctly handle enqueued receive
Integrating the active message shared memory broke enqueued receive
(on trylock/handoff).

This patch fixes the problem. The key idea here is that if
there is an unexpected message, let it complete first and complete
the request pre-allocated by CH4 separately, in contrast to the
current approach which simply tears down the unexpected message.

Tearing down the unexpected message is definitely not going to work
with the active message shared memory, because the shmmod may capture
the outstanding unexpected message internally and make progress on it.
So it shouldn't be torn down until it actually completes.