autogen.sh: patch libtool for detecting system lib search path
Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>

The /sbin/ldconfig now can correct report system library search path
which is more reliable than parsing the /etc/ld.so.conf file. We use the
check the results from ldconfig against the cached paths to determine
whether ldconfig is working properly. If not, we fallback to the
manual parsing of the ld.so.conf. In the fallback path, we will add
any missed cached path.