coll: refactor mpir_coll.h
This header declares all MPIR collective functions. It is a catalog of
all collective types and algorithms. The original files are cluttered
with parameter declarations, which are often long and repeats many
times. It is difficult to scan through the catalog and it is easy to
make inconsistent function signatures.

The refactor defines common parameters as macros (e.g. PARAMS_BCAST) --
one for each collective operation. These macros not only makes
parameters consistent and shortens code significantly in this file, it
also benefits many other files as the same parameter signature are often
needed in many internal functions.

The refactor also systematically orders all functions, first according
to the order in MPI standard, then follows `intra`, `inter` and
`allcomm` order.

Due to its regularity, such file is a potential target for script
generation -- infact, I generated this file using a template. This
refactoring provides foundation for further refactoring.