datatypes: get rid of DLOOP_DATALOOP_(ALL_BYTES|DEFAULT).
These are ancient flags that we used for devices that supported
heterogeneity vs. not.  We no longer support heterogeneity and we need
actual element information (not just bytes) at the lower levels to
take advantage of network operations anyway.  So we don't need these
flags anymore.  Also get rid of the corresponding flags in the
function declarations.

Signed-off-by: Min Si <msi@anl.gov>