ch4: Account for the existence of multiple netmod VCIs
- Store the total number of netmod VCIs in the global
MPIDI_CH4_Global struct.
- Ensure progress on all the netmod VCIs. For now, bindly progress
all the VCIs in a for loop. A smarter, more effective progress
function (one that reduces unnecessary progress) requires further
investigation.

Currently, the NM and SHM progress apis take the VCI as an int.
Hence, this commit treats the VCI as an int in the CH4 progress
function as well.

Signed-off-by: Rohit Zambre <rzambre@uci.edu>