ch4: Define the properties of a VCI
The enums define the various functionalites, types, and
properties of a VCI. Refer issue #3356 for details.

At the CH4 layer, the VCI is treated directly as an int. Hence,
there is no explicit type for a VCI. This is consistent with the
existing NM/SHM_progress APIs. The underlying netmod can map
integers to pointers if it needs to define the VCI as a different
type.

Signed-off-by: Rohit Zambre <rzambre@uci.edu>