ch4/netmod: Define the VCI pool APIs
The MPIDI_NM_vci_alloc API allows CH4 to request for a VCI from
the netmod's internal VCI pool. CH4 can request for VCIs with
certain functionalities and properties through the `resource`,
`type`, and `properties` arguments of the alloc API. If the netmod
is not able to satisfy CH4's request (returning vci as a negative
integer), CH4 can fallback to requesting a generic VCI.

The MPIDI_NM_vci_free API allows CH4 to return a VCI back into the
netmod's internal pool.

Refer issue #3356 for an example.

Signed-off-by: Rohit Zambre <rzambre@uci.edu>