datatypes: merge MPIR_Segment_alloc and MPIR_Segment_init
Now we no longer have an explicit MPIR_Segment_init function.  The
allocation function performs the necessary initialization too.

This is primarily done so that we can eventually hide the segment
structure from the lower layers.  Before this patch, since we had an
init function, the lower layers can declare the segment structure as a
local variable and initialize it.  That means the segment structure
will have to be fully exposed to the lower layers and cannot be hidden
by the dataloop layer (it is no longer opaque).

Providing only an allocation function means that the segment structure
is only accessed through a pointer by the lower layers, and thus can
be hidden by the dataloop layer.