test/mpi: Add xfail scripts
These scripts are to be used by Jenkins jobs. xfail.conf contains
information on which tests are expected to fail based on MPICH
configuration. This will allow bug fix PRs to include a commit to
re-enable any tests which are now expected to pass, rather than having
to bounce between repos.

Signed-off-by: Wesley Bland <wesley.bland@intel.com>