ch4/ucx: Implement netmod VCI pool with 1 VCI
A VCI in UCX is the UCP worker. The pool of VCIs in the UCX netmod
is simply an array of UCP workers. The rank to AV mapping will
return an array of UCP endpoints where each endpoint represents
a connection to one of the multiple workers present on the target
rank.

The pool consists of just one VCI for now; hence, we set the value
of int vci to 0 in many places. Once the CH4 utility function to
get a VCI is available, the netmod should use that function to get
the VCI required for the operation.

The VCI is also stored in the requests so that they can be obtained
during operations like mrecvs, cancel, etc.

TODO: need to implement proper business card exchange in
ucx_comm.h for the ROOTS_ONLY optimization

Signed-off-by: Rohit Zambre <rzambre@uci.edu>