ch4/ofi: MR key allocation rework
This commit simplified the MR key allocation in OFI. The OFI only
requires MR key to be unique. The key is used in RMA windows and the LMT
protocol. This patch does two things:

1. Rework the allocator to support 64bit keys that can be directly used
in OFI memory registration.

2. Using one global allocator for all MR keys. The key no longer have
complex layout of context ID, key type and unique number.

The rational of this patch are as follow.

1. There is no need to have separate allocator for window id and RMA id
in LMT. As the only requirement of uniqueness which can be guaranteed
with only on allocator. This also removes the need of having key type
field.

2. The per-comm allocator design is overcomplicating things with little
benefit. Since the key allocation only happens at window creation or
latency insensitive code path (large message), using per-comm allocator
offers little to none performance edge.

Author: Alexander Sannikov <alexander.sannikov@intel.com>
Co-Authored-by: Yanfei Guo <yguo@anl.gov>

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>