ch4/shm: define SHM internal control routines
POSIX is always enabled as basis of SHM, thus we extend the POSIX
progress engine to include SHM internal control header handling.

Changes in POSIX_am_header_t:
(1) add a 1-bit kind parameter to indicate whether it is a SHM ctrl
hdr or CH4 hdr (0 or 1);
(2) the handler_id is reduced to 7-bit (current max value is only 64).