ch4/fbox: Add fastbox polling cache
Add a small cache to keep track of fastboxes most likely to be full
(those where receives have already been posted). When checking for
messages in fastboxes, use this cache first to find a full box and only
when this list is exhausted, move on to the remainder of the fastboxes.

Initially, the CVAR used to size the cache will be set to 3 because that
will keep the entire array in a single cache line.

Signed-off-by: Min Si <msi@anl.gov>