ch4: Add new configuration options to choose VCI selection type
The --enable-ch4-vci-selection allows the user to choose the way
in which a VCI will be selected. The possible options are
comm-only, comm-rank, comm-tag, and comm-rank-tag. Note that all
the options involve the communicator. There are no options that
depend only on the rank or tag, or their combination. The reason
for this design simplification is that it is possible for a
communicators to have dedicated (MPIDI_VCI_EXCLUSIVE) VCIs and
so, a rank-only or tag-only hash would need to be aware of which
VCIs cannot be hashed to, which requires complicated bookkeeping.

The communicator will point to a subset of the total number of
VCIs available. The rank/tag will determine which VCI among the
subset will be used for the operation. In comm-only, all the
operations in a communicator will funnel through one VCI.

Signed-off-by: Rohit Zambre <rzambre@uci.edu>