ch4: Use the netmod VCI pool API during comm creation
During MPI_Init, CH4 allocates a generic, shared VCI and stores
it in the global data structure.

During any other communicator creation, CH4 first requests for
an exclusive VCI. If the request fails, it requests for a
generic, shared VCI.

The CH4 communicator object will store VCIs inside a hash region.
We do it this way to store the mapping between communicators and
VCIs that are dedicated to the communicator. Keeping the VCIs
external to the communicator would require the hash function
to know which VCIs it cannot hash to.

The `multi` struct of the union will store a list of VCIs and
the number of VCIs in the list. This struct will be used when
the selection of the VCI depends on the rank/tag in addition to
the communicator. However, if the selection depends only on the
communicator, the size of the list would be one and accessing
the VCI from the list would involve a pointer dereference. To
avoid this, the `single` struct of the union will be used when
it is guaranteed that all the operations of a communicator
will be funneled through the same VCI (as is the case with the
COMM_ONLY hash type).

Co-authored-by: Pavan Balaji <balaji@anl.gov>
Signed-off-by: Rohit Zambre <rzambre@uci.edu>