ch4: Add CH4 functions to get a VCI for pt2pt operations
The functions will return a VCI depending on the VCI selection
type. Note that if the user chooses to specify VCI selection
method at runtime, the overhead of a branch will exist since
the compiler cannot optimize out the switch statement.

Both the TX and RX operations will call these functions to obtain
the VCI to operate on just before executing the netmod operation.

Co-authored-by: Pavan Balaji <balaji@anl.gov>
Signed-off-by: Rohit Zambre <rzambre@uci.edu>