posix: integrate numa-aware shared memory segment allocation in fbox
This patch modifies the previous fbox segment allocation mechanism to
make it numa and heterogeneous memory-aware. This is done by counting
the number of available numa nodes used by MPI processes and creating
an equal number of shared memory segments (instead of just one). Each
of these segments will contain the fbox elements for the processes
located in the corresponding numa node and can be bound to the requested
type of memory (i.e., DDR or MCDRAM).