ch4: Rename progress hook counter to avoid semantic confusion
`active_progress_hooks` currently is incremented and decremented
whenever a progress hook is registered or deregistered,
respecrtively. The use of this counter is also to check if
there are any registered hooks in the first place before
performing any progress, not to check if the hook is active.
Hence, the current naming of this counter is confusing since a
registered hook is not an active hook.

The check for the hook's activeness is done through the `active`
field of the progress_hook_slot struct.

Signed-off-by: Pavan Balaji <balaji@anl.gov>