ch4/progress: Remove unnecessary iterations on global progress hooks
Currently, the loop to make progress on hooks registered globally
with CH4 iterates MAX_PROGRESS_HOOKS number of times even if the
number of registered hooks is lesser than MAX_PROGRESS_HOOKS.

This commit uses the value of the register_progress_hooks counter
to set the number of iterations of the loop that progresses the
globally registered hooks. This also reduces the number of times
the lock is taken to progress the hooks in the case of the
multi-thread models that are not `global`.

Signed-off-by: Pavan Balaji <balaji@anl.gov>