datatypes: cleanup public/private interfaces in dataloops/segments
segments and dataloops are internal type representation for datatypes
and should not be directly exposed to the other layers.  This patch is
a first step in this process.  It does the following things:

1. It refactors dataloop private functions, structures and variables
into the dataloop directory, so they are not exposed to the other
layers.

2. It updates local-only functions that are used in a single file to
be static functions.

3. It updates functions that are only used within the dataloop
directory to the MPII_ namespace.

4. It removes the DLOOP_ namespace, which is not an accepted namespace
within MPICH.

5. It removes incorrect usage of the MPIDU_ namespace in some cases.

There are no functionality changes in this patch.

No reviewer.