ch4/progress: Modularize progress with thread-safe functions
In terms of code design, this commit modularizes
MPIDI_Progress_test into progress of hooks, progress of the
netmod, and progress of the shmmod. Each of these modularized
functions are thread-safe.

In terms of performance, this commit adds an extra lock/unlock
pair on the `vci_lock`. However, this extra lock/unlock should
exist as an implication of generalizing the definition of the
VCI for both the netmod and the shmmod. Ideally, we should have
a separate `nm_vci_lock` and `shm_vci_lock` even in the case of
a single VCI. The current code simplifies this by just using
one lock for both NM and SHM VCIs. This commit maintains this
simplification.

Signed-off-by: <To be reviewed>