ch4: Account for the existence of multiple VCIs
- Store the total number of VCIs in the global MPIDI_CH4_Global
struct.

- The current MPIDI_Progress_test performs a global progress on
the hooks, netmod and shmmod. When multiple VCIs exist, a global
progress would need to progress all CH4 VCIs. Currently, the
net/shmmods implement only 1 VCI. Hence, this commit loops over
the number of VCIs returned from the net/shmmod. In the future,
the progress should loop over the number of VCIs allocated by
CH4, and not the number of VCIs in the net/shmmod pool.

Signed-off-by: <To-be-reviewed>