ch4/progress: Add infrastructure to choose global or per-VCI progress
This commit puts the current (global) progress code into a separate
global progress function, MPIDI_Progress_global. Similarly, it
introduces a new function for per-VCI progress.

MPIDI_Progress_test will now first check if it needs to perform a
global progress with the `progress_type` argument. If yes, it will
ignore the rest of the function arguments and call the global
progress function.

[Note that global progress is separate from MPIDI_PROGRESS_ALL. We
can have global progress without progressing certain types of hooks.
For example, when the netmod needs to call the progress function,
it could call global progress but it cannot progress
MPIDI_PROGRESS_HOOKS.]

If `progress_type` is MPIDI_PROGRESS_TYPE__VCI, then the progress
function will progress only one VCI that is pointed to by the
`vci` argument. The `vci_type` argument specifies if the VCI to
progress is a NM VCI or SHM VCI. However, it is possible that
progress on a certain VCI is dependent on the progress on another.
To account for this, we call global progress after
MAX_VCI_PROGRESS_ATTEMPTS on the single VCI. The
`progress_attempt_counter` is reset whenever global progress is
performed. The current value of MAX_VCI_PROGRESS_ATTEMPTS has been
chosen without any basis.

Additionally, this commit also puts the definitions of the hook
flags into an enum.

Signed-off-by: <To be reviewed>