ch4/progress: Remove mutual exclusion for registered-hooks-code
The counter for globally registered hooks is incremented and
decremented only during init and finalize time, respectively.
Hence, we don't need this counter to be an atomic one. For the
same reason, we don't need to take a lock during the registration
and deregistration of the hook.

In a sense, this commit explicitly defines the limits of the
usage of the `register/deregister` functions of the `MPID_Progress`
class. The limits are that the two functions cannot be used
concurrently (neither can the two threads call the same function,
nor can they call the different functions simultaneously). If, for
example, the registration of hooks could occur dynamically (not
statically during init time), then the MPID_Progress class would
need to expose thread-safe versions of these functions possibly
using the `safe` and `unsafe` suffixes in their names to
differentiate between the functions. Since, we don't need a
thread-safe version of this function today, this commit doesn't
add the `unsafe` suffix to the existing functions.

This commit also removes the check for `registered_progress_hooks`
being non-zero in the if condition. The for loop will not iterate
if there are zero hooks.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>