mpir: introduce hardware topology abstraction layer
Current comm_split_type implementation relies on hwloc/netloc direct
calls, which have to be guarded by ifdefs to allow builds that do not
use them to work. However, this solution is not ideal and makes the code
hard to maintain and prone to bugs.

This patch introduces a new hardware topology abstraction layer that can
be used by the MPIR layer. It allows to remove all the direct references
to hwloc/netloc currently present in the MPIR layer and move all the
hardware topology specific code under a simple, clean API.