coll: refactor mpir_coll.h
This header declares all MPIR collective functions. It is a catalog of
all collective types and algorithms. The original files are cluttered
with parameter declarations, which are often long and repeats many
times and spans multiple lines. It is very difficult to read and
difficult to spot inconsistencies.

This refactor defines common parameters as macros (e.g. PARAMS_BCAST) --
one for each collective operation. These macros not only makes
parameters consistent and shortens code significantly in this file, it
also benefits many other files as the same parameter signature are often
needed in many internal functions. As the macro names are limited to the
list MPI public collective functions, it should be safe to be exposed to
global namespace.

The refactor also systematically orders all functions, first according
to the order in MPI standard, then follows `intra`, `inter` and
`allcomm` order, as well as `coll`, `icoll`, `icoll_sched` order.

Due to its regularity, such file is a potential target for script
generation -- infact, I generated this file using a template. This
refactoring provides foundation for further refactoring.