testsuite: detach cmdtests.in from autoconf
We can run directly as `perl commands/cmdtests.pl` so we don't need
configure the perl path. We also can pass in options `-cxx`, `-f77` and
`-bindir=[path]` directly to the script.

By detaching it from the MPICH configure, it is more fit to be used as
cross-implementation test suite.