ch3: use flatten/unflatten instead of manual flattening
Get rid of the manual flattening of the datatype for RMA communication
and use the flatten/unflatten functions.  This allows us to hide the
details of the datatype representation from the lower layers.

As a side benefit, all of the datatype internal parameters do not need
to be stored in the ch3 code and is hidden in the flattened
representation.  This allows the ch3 code to be simpler with respect
to various header formats that are no longer needed.

Signed-off-by: Min Si <msi@anl.gov>