get rid of MPI_Aint casting macros
We had added a bunch of MPI_Aint to/from (void *) macros, and macros
to make sure that one can fit inside the other, during the IBM BG/P
days.  This was done because BG/P was a 32-bit machine but IBM wanted
to support a 64-bit filesystem, so they created a version where
MPI_Aint could be upgraded to 64-bit even though "void *" was just
32-bit.

Luckily, such architectures are behind us.  We can now safely assume
that "void *" is large enough for our file systems for a while.

Signed-off-by: Rob Latham <robl@mcs.anl.gov>