common/hcoll: Apply the `MPIDI_HCOLL` namespace
Currently the HCOLL API and MPICH's HCOLL code share the same
`hcoll_*` prefix, which is confusing.

This commit adds the `MPIDI_HCOLL` namespace to all functions
and variables related to the MPICH's HCOLL code.

In the process, the commit also renames some of the functions
defined in the MPIDI_HCOLL class so that they don't contain
numbers when unnecessary.

Signed-off-by: <To be reviewed>