comm_split_type: fix bug in node level split
When splitting communicator by subcomm_min_size or min_memsize the
current comm_split_type code does a first split at the network level and
then, if subcomm_min_size is still smaller than the number of processes
in the node, it further splits the subcomm at the node level.

This is done by grouping processes into fine grain subtrees and then
recursively merging them together until the desired comm size is
reached. The number of processes in each subtree is accounted for by an
array (processes_cpuset[]) which, right node, is statically set to one
in the code, thus assigning one process to every subtree.

This patch fixes this bug by properly incrementing the number of
processes.

Signed-off-by: Pavan Balaji <balaji@anl.gov>