ch4: fix zero-byte recv in unexpected AM recv routine
When an AM receive is matched and BUSY status is set (i.e., no matching
posted rreq at header arrival, thus the progress engine has to copy
data into a temporary buffer), and data is finally copied
into the recv buffer at MPIDIG_handle_unexp_cmpl if it is later matched
with a rreq. In this routine, although it is unlikely the data size is
zero, we reformat the code to support zero-size recv datatype.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>