ch4/vci: change how comm->vci mapping is done
Previously, we hash comm to a VCI id on the critical path, which is
costly. This patch embeds the VCI id as field in a comm object to allow
direct access. Hashing only occurs at comm creation time.

This patch also exposes control over the hashing function to be used as
a CVAR to simplify testing and tuning various hashing algorithms.