request: Fix bug in non-NULL request calculation
Both testany and waitany functions skip over inactive or NULL requests
before handing down to the device layer. However, the method for
discovering the first non-NULL request could erroneously skip the
first request in the array. To fix, we initialize the first non-NULL
request to an invalid index in the array (count) and set to valid
index later if one is found.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>