shm: Create topology aware intra_node trees
Creates collective specific trees which leverage the memory hierarchy of
a node. On a sample machine with multiple sockets and multiple cores per
socket, where each rank is bound to a core, a rank per socket is chosen
as the leader rank. For a bcast friendly tree, a socket leader interacts
with other socket leaders first, then with the ranks mapped on the
socket. For a reduce friendly tree, a socket leader interacts with ranks
within the same socket, the with other socket leaders. A left skewed
tree is created for bcast and right skewed tree (children are added in
the reverse order) for reduce.
Impact and benefits of these optimizations have been studied in the
Supercomputing 2018 paper - "Framework for scalable intra-node
collective operations using shared memory" by Jain et.al.
https://dl.acm.org/citation.cfm?id=3291695

Co-Authored-by: Surabhi Jain <surabhi.jain@intel.com>
Signed-off-by: Yanfei Guo <yguo@anl.gov>