ch4/progress: Remove unnecessary mutual exclusion for func_ptr
The `func_ptr` of a progress hook is set and unset only during the
hooks' registration and deregistration which occur during init and
finalize time. Hence, once we have read that a hook is active, we
do not need to protect the load of the `func_ptr` of the hook with
the `PROGRESS_HOOK_MUTEX`.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>