ch4/endpoints: assert that we don't allocate more than we can chew
This is a workaround. We should be able to multiplex endpoint handles to
physical resources. But currently this is needed because of the upcoming
AM support for endpoints that require maintaing a mapping table between
context ids + endpoint ids to comm objects. This table will be
preallocated with a predefined size; thus the assertion.