ch4/netmod: Declare the VNI pool API and define in stub
`MPIDI_NM_vni_alloc` allows CH4 to request for a VNI from the
netmod's internal pool. CH4 can request for a VNI with certain
resources and properties through the `resource`, and `properties`
arguments. If the netmod is not able to satisfy CH4's request
(returning vni as MPIDI_VNI_INVALID), CH4 can fallback to
requesting VNI with generic resources and properties.

`MPIDI_NM_vni_free` allows CH4 to return a VNI back to the
netmod's internal pool.

This commit defines the functions in the stub netmod.

Signed-off-by: <To be reviewed>