ch4/shm: Define a skeleton of the the VSI pool API
`MPIDI_SHM_vsi_alloc` allows CH4 to request for a VSI from the
shmmod's internal pool. CH4 can request for a VSI with certain
resources and properties through the `resource`, and `properties`
arguments. If the shmmod is not able to satisfy CH4's request
(returning vni as MPIDI_VSI_INVALID), CH4 can fallback to
requesting VSI with generic resources and properties.

`MPIDI_SHM_vsi_free` allows CH4 to return a VSI back to the
shmmod's internal pool.

This commit sets up the basic VSI management functions to
allocate and free VSIs in the shmmod's glue layer. We choose to
implement the VSI at the glue layer since, unlike netmods,
multiple shmmods can exist and a VSI will most likely have a
combination of resources from the multiple shmmods.

TODO: what is a VSI?

Signed-off-by: <To be reviewed>