ch3: move segment management one layer down the stack.
In the current code, the ch3 layer would create segments and pass them
down to the netmod or shared memory functions.  This would force the
netmods/shmmods to use segments and not benefit from any potential
optimizations that the hardware provides.  This patch moves the
segment allocation code one layer down.  This way, the netmods/shmmods
can choose to use either segments or something else.