ch3: get rid of buf_off variable.
We are overcomplicating the copy from one noncontiguous buffer to
another by trying to accommodate for the case where we pack X bytes,
but are unable to unpack those X bytes into an equivalent
noncontiguous buffer.  However, this case should not be possible, in
practice.  Specifically, the pack/unpack routines should always limit
how much data they copy to keep it atomic with respect to an MPI basic
datatype.  In practice, we keep the copies atomic to C datatypes (so
pairtypes are not atomic).  Either way, MPI requires the source and
destination datatype signatures to match, so the granularity of
packing/unpacking should always be the same.  Hence, this offset is
not needed.

Signed-off-by: Giuseppe Congiu <gcongiu@anl.gov>