ch4: Allocate/free VCIs during init/free and Comm create/free
During MPI_Init, CH4 allocates a shared VCI with generic resources
and properties. This is the root VCI. If any future VCI allocation
fails, CH4 will revert to using the root VCI.

During any new communicator creation, CH4 first requests for an
exclusive VCI. If the request fails, it requests for a shared VCI
with generic resources and properties. (In the current
implementation of the pool, we maintain only one shared VCI which
is the root VCI.) The CH4 communicator will store the index of the
VCI allocated to it inside a hash region.

Co-authored-by: Pavan Balaji <balaji@anl.gov>
Signed-off-by: <To be reviewed>