ch4/progress: Remove counter that keeps track of state for non-global critical sections
This global counter is a point of high contention when multiple
threads are making progress. Since we will use a different
progress mechanism to detect whether or not progress has been
made when multiple VCIs are involved, this commit gets rid of this
counter on the critical path for the non-global critical sections.