logging: disable logging options
The rlog build is broken. In addition, its autogen script
`maint/genstates` relies on `FUNCNAME` macros, which are being removed
in a coming patch. Having the script active in autogen will emit tons of
warning noise. It is also a bad habit to leave broken options active.

This PR changes autogen to not to run maint/genstates by default. It
moves logging/timing related code from configure.ac to
src/util/logging/subconfigure.m4 and disabled by default.

Having timing and detailed logging is potentially useful, which is the
reason this PR choose to disable (and refactor) rather than remove. We
need to decide either to fix or to remove in future.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>