test: minimize difference of `FROM_MPICH` configure
Currently testsuite is being configured inside the main mpich subdir
configure, with `FROM_MPICH` set. This flag on one hand short-cuts many
configure options to avoid compile tests; however, on the other hand,
its behavior become very tied to MPICH and makes some behavior
mysterious.

Since one of the goal of testsuite is to develop a "MPI" testsuite,
rather than a "MPICH" testsuite. While we want to retain some of the
MPICH specific tests (such as MPI extensions), the rest of them should
test MPICH as if it is any standard compliant MPI implementation.
Therefore, we should minimize the MPICH specific configure as much as we
can. It should also simplify the configure scripts significantly as we
don't need maintain two configure paths -- one for MPICH and one for
others.