ch3: reset recv segment when restart with SRBuf unpack
When receiving partial data with derived datatype, ch3 first tries
to generate iovs and let the progress engine copy incoming data
into recv buffer. However, if the datatype is too complex,
MPL_IOV_LIMIT count of iovs cannot represent the entire datatype.
Then we have to use a SRBuf to temporarily store received data
(packed), and unpack it after all data arrives.

The original code has a bug that it uses the segment_ptr to try out
the iov approach but does not reset it when switching to SRBuf unpack.
Thus, the datatype representation was incomplete at unpacking time. The
error can be reproduced with ch3/sock and ch3/tcp which may receive
partial data. This patch fixes it by free-alloc the segment_ptr when
switching to SRBuf approach.

Fixes pmodels/mpich#3755

Signed-off-by: Pavan Balaji <balaji@anl.gov>