ch4: Define the VCI, the VCI pool and its management functions
VCI - Virtual Communication Interface is CH4's communication
object that has various types (MPIDI_vci_type_t), resources
(MPIDI_vci_resource_t), and properties (MPIDI_vci_property_t).

Each VCI has its own lock, a lightweight request for
operations that complete immediately on the VCI, a VNI, and
a VSI.

CH4 maintains a pool of VCIs that is created during init time.
This pool is simply an array of VCIs, the memory for which is
allocated dynamically to accommodate for the maximum number of
VCIs possible. VCIs from this pool will be dynamically allocated
to objects like communicators during communicator creation time.

Since multiple communicators can be created simultaneously, the
VCI pool has its own lock that is taken to provide mutual
exclusion during the allocation or deallocation (free) for a VCI.
This lock also maintains the thread safety of the
allocation/deallocation of VNIs and VSIs since the VNIs and VSIs
are allocated/freed only when a VCI is allocated/freed.

The CH4 request also now stores the VCI that was used for the
request.