common: Apply new naming rules to the Sched class
This commit introduces the following class-style naming to
the Sched objects and functions:

- Sched_element*: related to a schedule (element is a
schedule)
- Sched_list*: related to the global list of schedules

We apply this new naming list so that the purpose and the
class of objects they work on is clear.

This commit also adds the `MPIDU` namespace to the global
`all_schedules` list.

This commit defines the MPIR_Sched_list_t and simply points
it to the existing `MPIDU_all_schedules` global list. In this
process, the type of this global list has been renamed from
`MPIDU_Sched_state` to `MPIDU_Sched_list` to keep the names
similar between the `MPIDU_Sched*` and `MPIR_Sched*`
objects/functions. Once issue #3596 is resolved, this global list
will sit in src/include/mpir_sched.h.

Signed-off-by: <To be reviewed>