coll: Apply new naming rules to TSP code
This commit introduces the
`[MPIR_TSP|MPII]_[Genutil|Stubutil]_queue_*` naming for all
functions whose purpose is to deal with the global queue of
collective-schedules.

In the process, the commit renames the
`Gentran_scheds_are_pending` function to
`Genutil_queue_has_pending_scheds`, replacing its namespace from
`Gentran` to `Genutil`. This maintains consistency between the
naming of the functions.

Additionally, this commit gives the global `coll_queue` the `MPII`
namespace.

Signed-off-by: <To-be-reviewed>