ch4: Prevent premature completion of SHM request
When canceling a SHM request on MPI_ANY_SOURCE receive,
MPIDI_SHM_mpi_cancel_recv may complete the request, but we do not
want it to be visible until its partner's status is copied to the
SHM request.
This is important when an asynchronous progress is on, because
a user thread may immediately see the (incomplete) status.

This patch bumps the request refcount by one to prevent request completion inside
MPIDI_SHM_mpi_cancel_recv.