conf: Move HAVE_HWLOC to CH4 and name BUILD_TOPOTREES
The (case-sensitive) automake macro HAVE_HWLOC was previously added in
the top level configure.ac to be able to detect when HWLOC was enabled
to help determine when to build topology-aware trees in CH4.

However, just knowing if HWLOC was available was actually not enough.
The Linux-only header, sched.h, is also needed to be able to use the
needed features in HWLOC. This became apparent when trying to build on
MacOS, which doesn't have this header.

So this patch moves the check down to CH4 (because that's the only place
where it's used) and adds a check to see if the needed features of
sched.h are available. Thus the new macro is called BUILD_TOPOTREES
instead of HAVE_HWLOC.

Then, this patch changes the ch4/shm makefile to check this macro before
building the topology aware trees code.