warnings: only add alignment attribute if necessary
OFI providers may require bigger alignment for `stuct iovec iov`, in
which case we need enforce the alignment. In the case of natural
alignment is already sufficient, it doesn't seem that we need *reduce*
the alignment. In fact, `icc` ver 2019 will warn in the case of
alignment reduction -- unless the struct is packed (which is not), the
reduction in alignment will be ignored.

The `struct iovec`'s natural alignment is at pointer size -- 8 in 64bit
case. The `MPIDI_OFI_IOVEC_ALIGN` defined in
`src/mpid/ch4/netmod/ofi/ofi_capability_sets.h` is set to 1 in all
cases, which means in most cases, we don't need additional alignment
overwrite or it will be an alignment reduction.

This patch removes the alignment attribute in case of reduction.

Signed-off-by: Wesley Bland <wesley.bland@intel.com>