warnings: icc: enumerated type mixed with int
Compilers have no way to check an int variable against the enumerated
values. Thus, having an variable in enum type gains no benefit.
If it is ever going to be used as an int, e.g. arithemetics, bit and,
bit or, just use int. The name of the variable should provide sufficient
hint.

Signed-off-by: Wesley Bland <wesley.bland@intel.com>