ucx: add ucp_rkey_destroy in MPIDI_UCX_mpi_win_free_hook
We are supposed to destroy any rkeys allocated with
`ucp_ep_rkey_unpack`. Since ucx internally try to use `mpool` for
limited allocation, it may not always leak memory, but ucx will warn
about it -- unless we supress with UCX_LOG_LEVEL=error, which we do in
our Jenkins tests.

Signed-off-by: Pavan Balaji <balaji@mcs.anl.gov>