test: disable gather_big on all platforms.
The gather_big test requires 1GB of send buffer on each process and 8
GB of receive buffer on the root process.  When running on a single
node, that'd mean we need 16GB of buffer space just for the user
buffers.  On top of that, we use a binomial algorithm within MPICH for
large Gathers, and each intermediate process allocates temporary
buffers for the entire incoming message.  That's an additional 1GB for
4 processes, and 3GB for 2 processes.  That'd put us at a total of
24GB memory requirement.  That's the maximum amount of memory
available on our nodes, making this test somewhat unreliable.

Also see issue#3770.