mpi: get rid of buf_off variable.
This patch is similar to [7305083b4cd2], but is for the MPIR-level
localcopy function.  We are overcomplicating the copy from one
noncontiguous buffer to another by trying to accommodate for the case
where we pack X bytes, but are unable to unpack those X bytes into an
equivalent noncontiguous buffer.  This patch removes that logic as it
is not needed.