datatypes: generalize pack/unpack routines.
The current MPIR_Pack|Unpack_impl routines are too specific to
MPI_Pack|Unpack, and are not suitable for how the rest of MPICH uses
it.  For example, the code assumes that the out buffer fits the entire
input.  Also, there's no ability to provide an offset for the input
buffer from where to start.

In this patch, we generalize these routines that provides these
capabilities.  The new routines drop the position argument, which is
really an artifact of Fortran usage, and is not really necessary in
the C code where the pointer can be incremented.