datatypes: add a shortcut for datatype seek
If the user uses a fresh segment and is trying to pack/unpack at an
offset, the segment manipulate code requires the segment pointer to be
"seeked" to the correct location.  This patch adds a shortcut for the
case where we are only performing the seek and not the actual
pack/unpack.