ch3: nontrivial replacements of segment_pack with pack_impl
These would theoretically add an additional overhead because we are
allocating a segment for each pack operation.  However, from the
measurements, we seem to be adding 3-4 microseconds for each
allocation, while the pack size is for 256KB.  Plus, there are data
transmission overheads involved too.  Overall, the percentage cost
should be fairly small with this change.