maint: code-cleanup.sh fixups
Some refactors plus accepting mutliple files and filter them through
ignore list.

Let the code cleanup script safely take any filename and it will only
try to format it if the file is one we actually want to format.
Previously, this worked when using --all or --recursive, but not with
particular filenames. This is a problem if you run the cleanup script
with a long list of files, some of which shouldn't be touched.

Co-authored-by: Wesley Bland <wesley.bland@intel.com>
Co-authored-by: Hui Zhou <hzhou321@anl.gov>