init: Workaround Intel compiler TLS bug on macOS
Work around a bug in the Intel compiler on macOS that does not treat
MPIR_Per_thread as thread-local storage when it is
instantiated. References to MPIR_Per_thread in other files would then
conflict with the non-TLS version. Add an explicit reference in
initthread.c to trick the compiler. Fixes pmodels/mpich#3437.