coll: Remove MPID_I<coll>_sched functions
We already provide an device-override point for each collective in the
MPID_I<call> functions.  A second override point is unnecessary.

Removing this override also has the advantage of discouraging devices
from implementing sched-function overrides instead of the
request-based function overrides.  The intent is to eventually delete
the sched-based functions.