coll: move algorithm selection into a single function.
This allows a single switch statement to deal with both gentran and
sched-based algorithms.  The upside of this refactoring is that all
selection is in one place making it easier for future addition of
other selection models, such as json-based selection.  The downside is
that when someone wants to select from only the sched-based
algorithms, they always get the _auto functions.  No overriding based
on CVAR is possible in such cases.

The following algorithms are affected in performance because of this
change --

Intra communicator algorithms:
 - MPIR_Ibcast_intra_scatter_recursive_doubling_allgather
 - MPIR_Iallreduce_intra_smp
 - MPIR_Ibcast_intra_smp
 - MPIR_Ireduce_intra_smp
 - MPIR_Iscan_intra_smp

Inter communicator algorithms:
 - MPIR_Iallgather_inter_local_gather_remote_bcast
 - MPIR_Iallgatherv_inter_remote_gather_local_bcast
 - MPIR_Iallreduce_inter_remote_reduce_local_bcast
 - MPIR_Ibarrier_inter_bcast
 - MPIR_Ibcast_inter_flat
 - MPIR_Igather_inter_short
 - MPIR_Ireduce_inter_local_reduce_remote_send
 - MPIR_Ireduce_scatter_inter_remote_reduce_local_scatterv
 - MPIR_Ireduce_scatter_block_inter_remote_reduce_local_scatterv
 - MPIR_Iscatter_inter_remote_send_local_scatter